Implementation/requirements.txt
torchvision
Shapely
tifffile